OLD_AI_AGENT_FILES/Custom_Ingredient_LLM/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai>=1.54.3
aiohttp==3.10.10
python-dotenv==1.0.1
pydantic>=2.11.0